requirements.txt
openai==1.30.1
httpx==0.27.0
python-docx==1.2.0
PyPDF2==3.0.1
python-dotenv==1.2.2